chore: added basic documentation and some project refactoring
